@@ -3,15 +3,13 @@
   "version": "0.1.0",
   "private": true,
   "dependencies": {
-    "eslint-plugin-flowtype": "^3.13.0",
-    "prop-types": "^15.7.2",
-    "react": "^16.8.6",
-    "react-animated-burgers": "^1.2.7",
-    "react-dom": "^16.8.6",
-    "react-router-dom": "^5.0.0",
-    "react-scripts": "^3.4.0",
-    "sass": "^1.26.3",
-    "styled-components": "^5.0.1"
+    "prop-types": "^15.8.1",
+    "react": "^18.2.0",
+    "react-dom": "^18.2.0",
+    "react-router-dom": "^6.22.3",
+    "react-scripts": "^5.0.1",
+    "sass": "^1.75.0",
+    "styled-components": "^6.1.8"
   },
   "scripts": {
     "start": "react-scripts start",
