@@ -5,16 +5,16 @@ import Contact from "../Contact/index";
 import "./App.css";
 import Projects from "../Projects/index";
 import Experience from "../Experience";
-import { Route, Switch } from "react-router-dom";
+import { Route, Routes } from "react-router-dom";
 
 const App = () => (
   <main>
-    <Switch>
-      <Route exact path="/" component={MainMenu} />
-      <Route path="/experience" component={Experience} />
-      <Route path="/contact" component={Contact} />
-      <Route path="/projects" component={Projects} />
-    </Switch>
+    <Routes>
+      <Route exact path="/" element={<MainMenu />} />
+      <Route path="/experience" element={<Experience />} />
+      <Route path="/contact" element={<Contact />} />
+      <Route path="/projects" element={<Projects />} />
+    </Routes>
   </main>
 );
 
